Combine DOM-related concepts in Layout 2020 into dom.rs
