--- components/layout_2020/opaque_node.rs
@@ -1,17 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-use libc::c_void;
-use script_traits::UntrustedNodeAddress;
-use style::dom::OpaqueNode;
-
-pub trait OpaqueNodeMethods {
-    fn to_untrusted_node_address(&self) -> UntrustedNodeAddress;
-}
-
-impl OpaqueNodeMethods for OpaqueNode {
-    fn to_untrusted_node_address(&self) -> UntrustedNodeAddress {
-        UntrustedNodeAddress(self.0 as *const c_void)
-    }
-}
